Various changes to prep for 1.4.5
Removed old dev_support page, simplified the (now on-by-default) dev
support options. Added link to NM to "Upcoming events" widget. Replaced
old NM with new one. Randomized ALT text on dev support links based on
the day on which the meetup event takes place. Note that all this is
untested for this commit.